修改套餐画像接口  和  查看套餐详情接口

+++ yzl-product-common/src/main/java/cn/net/yzl/product/model/db/MealProduct.java
@@ -15,7 +15,7 @@ public class MealProduct {
     private Integer id;
 
     @ApiModelProperty(value = "套餐编码",name = "mealNo")
-    private Integer mealNo;
+    private Long mealNo;
 
     @ApiModelProperty(value = "商品编码",name = "productCode")
     private String productCode;
